+++ performance/bwt_search.c
@@ -8,13 +8,13 @@
 #include <string.h>
 #include <assert.h>
 
-static char *build_random(size_t size)
+static char *build_random(uint32_t size)
 {
     const char *alphabet = "ACGT";
     int n = strlen(alphabet);
     char *s = malloc(size + 1);
     
-    for (size_t i = 0; i < size; ++i) {
+    for (uint32_t i = 0; i < size; ++i) {
         s[i] = alphabet[rand() % n];
     }
     s[size] = '\0';
@@ -22,20 +22,20 @@ static char *build_random(size_t size)
     return s;
 }
 
-static char *sample_string(const char *string, size_t n, size_t m)
+static char *sample_string(const char *string, uint32_t n, uint32_t m)
 {
-    size_t offset = rand() % (n - m);
+    uint32_t offset = rand() % (n - m);
     char *p = str_copy_n(string + offset, m);
     return p;
 }
 
-static void edit_string(char *string, struct bwt_table *bwt_table, size_t m, float err)
+static void edit_string(char *string, struct bwt_table *bwt_table, uint32_t m, float err)
 {
     // I don't know if I need to test indels as well but it is
     // easier to simply sample substitutions so that is what I do.
     
     
-    for (size_t i = 0; i < m; ++i) {
+    for (uint32_t i = 0; i < m; ++i) {
         double p = rand() / RAND_MAX;
         if (p < err) { // edit 5% of the characters
             char a = rand() % bwt_table->remap_table->alphabet_size;
@@ -53,7 +53,7 @@ static void search(struct bwt_table *bwt_table, const char *p, int edits)
     init_bwt_approx_iter(&iter, bwt_table, p, edits);
     while (next_bwt_approx_match(&iter, &match)) {
         // do nothing
-        printf("match at %lu\n", match.position);
+        printf("match at %u\n", match.position);
     }
     dealloc_bwt_approx_iter(&iter);
 }
@@ -109,7 +109,7 @@ static const char *cigar_alignment(const char *cigar, const char *pattern,
 }
  */
 
-static unsigned long get_performance(struct bwt_table *bwt_table, size_t m, float err, int edits)
+static unsigned long get_performance(struct bwt_table *bwt_table, uint32_t m, float err, int edits)
 {
     assert(m > 0);
     
@@ -117,7 +117,7 @@ static unsigned long get_performance(struct bwt_table *bwt_table, size_t m, floa
     char *p;
 
     search_begin = clock();
-    for (size_t j = 0; j < 10; ++j) {
+    for (uint32_t j = 0; j < 10; ++j) {
         p = sample_string(bwt_table->sa->string,
                         bwt_table->sa->length,
                         m);
@@ -144,7 +144,7 @@ int main(int argc, const char **argv)
     
     clock_t time;
     
-    size_t size = 10000;
+    uint32_t size = 10000;
     s = build_random(size);
     
     init_remap_table(&remap_table, s);
@@ -162,15 +162,15 @@ int main(int argc, const char **argv)
     
     
 #if 1 // for comparison
-    for (size_t m = 50; m < 200; m += 50) {
+    for (uint32_t m = 50; m < 200; m += 50) {
         for (int edits = 1; edits < 4; ++edits) {
             float p = 0.05;
             while (p < 0.5) {
-                for (size_t rep = 0; rep < 5; ++rep) {
+                for (uint32_t rep = 0; rep < 5; ++rep) {
                     time = get_performance(&bwt_table, m, p, edits);
-                    printf("BWT-without-D %zu %f %d %lu\n", m, p, edits, time);
+                    printf("BWT-without-D %u %f %d %lu\n", m, p, edits, time);
                     time = get_performance(&bwt_table_D, m, p, edits);
-                    printf("BWT-with-D %zu %f %d %lu\n", m, p, edits, time);
+                    printf("BWT-with-D %u %f %d %lu\n", m, p, edits, time);
                 }
                 p += 0.05;
             }
@@ -179,7 +179,7 @@ int main(int argc, const char **argv)
     
 #else // for profiling
     
-    for (size_t n = 0; n < 1000; n += 100) {
+    for (uint32_t n = 0; n < 1000; n += 100) {
         for (int rep = 0; rep < 10; ++rep) {
             get_performance(n);
         }
